python/trade_bot/requirements.txt
 mysql-connector-python
 alpaca_trade_api
 mypy